added upadintg profile picture feture latest

+++ controllers/userController.js
@@ -78,7 +78,7 @@ module.exports.update = async function(req,res){
 
             
 
-            if(user.avatar){
+            if(user.avatar && fs.existsSync(path.join(__dirname,'..', user.avatar))){
                 fs.unlinkSync(path.join(__dirname,'..', user.avatar));
             }
